requirements.txt
# LiteLLM Config Generator - Developed by acidvegas (https://github.com/acidvegas)
# requirements.txt

apv
python-dotenv
litellm
litellm[proxy]
pydantic[email]